requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
discord.py==2.3.1
frozenlist==1.3.3
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreloadhttpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
multidict==6.0.4
python-decouple==3.8
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
urllib3==2.0.3
yarl==1.9.2